refactor: refactoring interceptor
