+++ src/main/java/gunnu/stocknote/auth/Auth.java
@@ -1,0 +1,14 @@
+package gunnu.stocknote.auth;
+
+import gunnu.stocknote.user.entity.UserRole;
+import java.lang.annotation.ElementType;
+import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
+import java.lang.annotation.Target;
+
+@Target(ElementType.METHOD)
+@Retention(RetentionPolicy.RUNTIME)
+public @interface Auth {
+
+    UserRole role() default UserRole.USER;
+}
