+++ src/main/java/gunnu/stocknote/stocknote/service/StocknoteService.java
@@ -2,7 +2,6 @@ package gunnu.stocknote.stocknote.service;
 
 import gunnu.stocknote.stocknote.dto.response.StocknoteResponseDTO;
 import gunnu.stocknote.stocknote.entity.Stocknote;
-import gunnu.stocknote.stocknote.repository.StocknoteJpaRepository;
 import gunnu.stocknote.stocknote.repository.StocknoteRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
@@ -12,8 +11,6 @@ import org.springframework.stereotype.Service;
 public class StocknoteService {
 
     private final StocknoteRepository stocknoteRepository;
-    private final StocknoteJpaRepository jpaRepository;
-
 
     public StocknoteResponseDTO createStocknote(
         final String title,
